Address beta feature review feedback
