+++ Sources/App/WorkspaceRuntimeSettings.swift
@@ -116,12 +116,12 @@ enum RightSidebarBetaFeatureSettings {
     static let defaultFeedEnabled = false
     static let defaultDockEnabled = false
 
-    static func isFeedEnabled(defaults: UserDefaults = .standard) -> Bool {
+    nonisolated static func isFeedEnabled(defaults: UserDefaults = .standard) -> Bool {
         guard defaults.object(forKey: feedEnabledKey) != nil else { return defaultFeedEnabled }
         return defaults.bool(forKey: feedEnabledKey)
     }
 
-    static func isDockEnabled(defaults: UserDefaults = .standard) -> Bool {
+    nonisolated static func isDockEnabled(defaults: UserDefaults = .standard) -> Bool {
         guard defaults.object(forKey: dockEnabledKey) != nil else { return defaultDockEnabled }
         return defaults.bool(forKey: dockEnabledKey)
     }
