+++ Sources/MainWindowFocusTypes.swift
@@ -1,0 +1,29 @@
+import Foundation
+
+struct FeedFocusSnapshot: Equatable {
+    var selectedItemId: UUID?
+    var isKeyboardActive: Bool
+
+    init(selectedItemId: UUID? = nil, isKeyboardActive: Bool = false) {
+        self.selectedItemId = selectedItemId
+        self.isKeyboardActive = isKeyboardActive
+    }
+}
+
+protocol FeedKeyboardFocusResponder: AnyObject {}
+
+enum MainWindowKeyboardFocusIntent: Equatable {
+    case mainPanel(workspaceId: UUID, panelId: UUID)
+    case rightSidebar(mode: RightSidebarMode)
+}
+
+enum MainWindowFocusToggleDestination: Equatable {
+    case terminal
+    case rightSidebar
+}
+
+enum MainWindowFindShortcutTarget: Equatable {
+    case mainPanelFind
+    case rightSidebarFileSearch
+    case none
+}
