+++ Sources/RightSidebarMode+Availability.swift
@@ -1,0 +1,32 @@
+import Foundation
+
+extension RightSidebarMode {
+    static func availableModes(defaults: UserDefaults = .standard) -> [RightSidebarMode] {
+        availableModes(
+            feedEnabled: RightSidebarBetaFeatureSettings.isFeedEnabled(defaults: defaults),
+            dockEnabled: RightSidebarBetaFeatureSettings.isDockEnabled(defaults: defaults)
+        )
+    }
+
+    static func availableModes(feedEnabled: Bool, dockEnabled: Bool) -> [RightSidebarMode] {
+        allCases.filter { $0.isAvailable(feedEnabled: feedEnabled, dockEnabled: dockEnabled) }
+    }
+
+    func isAvailable(defaults: UserDefaults = .standard) -> Bool {
+        isAvailable(
+            feedEnabled: RightSidebarBetaFeatureSettings.isFeedEnabled(defaults: defaults),
+            dockEnabled: RightSidebarBetaFeatureSettings.isDockEnabled(defaults: defaults)
+        )
+    }
+
+    func isAvailable(feedEnabled: Bool, dockEnabled: Bool) -> Bool {
+        switch self {
+        case .files, .find, .sessions:
+            return true
+        case .feed:
+            return feedEnabled
+        case .dock:
+            return dockEnabled
+        }
+    }
+}
