+++ cmuxTests/FileExplorerStateModePersistenceTests.swift
@@ -1,0 +1,85 @@
+import Foundation
+import XCTest
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+#elseif canImport(cmux)
+@testable import cmux
+#endif
+
+final class FileExplorerStateModePersistenceTests: XCTestCase {
+    private let modeKey = "rightSidebar.mode"
+    private let feedEnabledKey = RightSidebarBetaFeatureSettings.feedEnabledKey
+    private let dockEnabledKey = RightSidebarBetaFeatureSettings.dockEnabledKey
+
+    func testDisabledFeedStoredModeFallsBackToFiles() {
+        withSavedRightSidebarModeDefaults {
+            let defaults = UserDefaults.standard
+            defaults.set(RightSidebarMode.feed.rawValue, forKey: modeKey)
+            defaults.set(false, forKey: feedEnabledKey)
+
+            let state = FileExplorerState()
+
+            XCTAssertEqual(state.mode, .files)
+            XCTAssertEqual(defaults.string(forKey: modeKey), RightSidebarMode.files.rawValue)
+        }
+    }
+
+    func testEnabledFeedStoredModeSurvives() {
+        withSavedRightSidebarModeDefaults {
+            let defaults = UserDefaults.standard
+            defaults.set(RightSidebarMode.feed.rawValue, forKey: modeKey)
+            defaults.set(true, forKey: feedEnabledKey)
+
+            let state = FileExplorerState()
+
+            XCTAssertEqual(state.mode, .feed)
+            XCTAssertEqual(defaults.string(forKey: modeKey), RightSidebarMode.feed.rawValue)
+        }
+    }
+
+    func testModeSetterClampsUnavailableBetaModes() {
+        withSavedRightSidebarModeDefaults {
+            let defaults = UserDefaults.standard
+            defaults.set(false, forKey: feedEnabledKey)
+            defaults.set(false, forKey: dockEnabledKey)
+            let state = FileExplorerState()
+
+            state.mode = .feed
+            XCTAssertEqual(state.mode, .files)
+            XCTAssertEqual(defaults.string(forKey: modeKey), RightSidebarMode.files.rawValue)
+
+            defaults.set(true, forKey: dockEnabledKey)
+            state.mode = .dock
+            XCTAssertEqual(state.mode, .dock)
+            XCTAssertEqual(defaults.string(forKey: modeKey), RightSidebarMode.dock.rawValue)
+
+            defaults.set(false, forKey: dockEnabledKey)
+            state.refreshModeAvailability()
+            XCTAssertEqual(state.mode, .files)
+            XCTAssertEqual(defaults.string(forKey: modeKey), RightSidebarMode.files.rawValue)
+        }
+    }
+
+    private func withSavedRightSidebarModeDefaults(_ body: () -> Void) {
+        let defaults = UserDefaults.standard
+        let previousMode = defaults.object(forKey: modeKey)
+        let previousFeedEnabled = defaults.object(forKey: feedEnabledKey)
+        let previousDockEnabled = defaults.object(forKey: dockEnabledKey)
+        defer {
+            restore(previousMode, forKey: modeKey)
+            restore(previousFeedEnabled, forKey: feedEnabledKey)
+            restore(previousDockEnabled, forKey: dockEnabledKey)
+        }
+        body()
+    }
+
+    private func restore(_ value: Any?, forKey key: String) {
+        let defaults = UserDefaults.standard
+        if let value {
+            defaults.set(value, forKey: key)
+        } else {
+            defaults.removeObject(forKey: key)
+        }
+    }
+}
