+++ cmuxTests/RightSidebarCommandPaletteTests.swift
@@ -8,41 +8,74 @@ import XCTest
 #endif
 
 final class RightSidebarCommandPaletteTests: XCTestCase {
-    func testCommandPaletteIncludesActionForEachRightSidebarMode() throws {
-        let contributions = ContentView.commandPaletteRightSidebarModeCommandContributions()
-        let contributionsByID = Dictionary(uniqueKeysWithValues: contributions.map { ($0.commandId, $0) })
-        let context = ContentView.CommandPaletteContextSnapshot()
+    func testCommandPaletteIncludesDefaultRightSidebarModes() throws {
+        try withSavedBetaFeatureDefaults {
+            let defaults = UserDefaults.standard
+            defaults.removeObject(forKey: RightSidebarBetaFeatureSettings.feedEnabledKey)
+            defaults.removeObject(forKey: RightSidebarBetaFeatureSettings.dockEnabledKey)
+            let contributions = ContentView.commandPaletteRightSidebarModeCommandContributions()
+            let contributionsByID = Dictionary(uniqueKeysWithValues: contributions.map { ($0.commandId, $0) })
+            let context = ContentView.CommandPaletteContextSnapshot()
 
-        for mode in RightSidebarMode.allCases {
-            let commandID = ContentView.commandPaletteRightSidebarModeCommandID(mode)
-            let contribution = try XCTUnwrap(
-                contributionsByID[commandID],
-                "Expected command palette contribution for \(mode.rawValue)"
-            )
+            for mode in RightSidebarMode.availableModes() {
+                let commandID = ContentView.commandPaletteRightSidebarModeCommandID(mode)
+                let contribution = try XCTUnwrap(
+                    contributionsByID[commandID],
+                    "Expected command palette contribution for \(mode.rawValue)"
+                )
 
-            XCTAssertEqual(contribution.title(context), mode.shortcutAction.label)
-            XCTAssertEqual(
-                contribution.subtitle(context),
-                String(localized: "command.rightSidebarMode.subtitle", defaultValue: "Right Sidebar")
-            )
-            XCTAssertTrue(contribution.keywords.contains("right"))
-            XCTAssertTrue(contribution.keywords.contains("sidebar"))
-            XCTAssertTrue(contribution.keywords.contains(mode.rawValue))
-            XCTAssertTrue(contribution.when(context))
-            XCTAssertTrue(contribution.enablement(context))
+                XCTAssertEqual(contribution.title(context), mode.shortcutAction.label)
+                XCTAssertEqual(
+                    contribution.subtitle(context),
+                    String(localized: "command.rightSidebarMode.subtitle", defaultValue: "Right Sidebar")
+                )
+                XCTAssertTrue(contribution.keywords.contains("right"))
+                XCTAssertTrue(contribution.keywords.contains("sidebar"))
+                XCTAssertTrue(contribution.keywords.contains(mode.rawValue))
+                XCTAssertTrue(contribution.when(context))
+                XCTAssertTrue(contribution.enablement(context))
+            }
+
+            XCTAssertEqual(contributions.count, 3)
+            XCTAssertNil(contributionsByID[ContentView.commandPaletteRightSidebarModeCommandID(.feed)])
+            XCTAssertNil(contributionsByID[ContentView.commandPaletteRightSidebarModeCommandID(.dock)])
         }
-
-        XCTAssertEqual(contributions.count, RightSidebarMode.allCases.count)
     }
 
     func testCommandPaletteRightSidebarActionsUseModeShortcutActions() {
-        for mode in RightSidebarMode.allCases {
-            XCTAssertEqual(
-                ContentView.commandPaletteShortcutAction(
-                    forCommandID: ContentView.commandPaletteRightSidebarModeCommandID(mode)
-                ),
-                mode.shortcutAction
-            )
+        withSavedBetaFeatureDefaults {
+            let defaults = UserDefaults.standard
+            defaults.set(true, forKey: RightSidebarBetaFeatureSettings.feedEnabledKey)
+            defaults.set(true, forKey: RightSidebarBetaFeatureSettings.dockEnabledKey)
+
+            for mode in RightSidebarMode.allCases {
+                XCTAssertEqual(
+                    ContentView.commandPaletteShortcutAction(
+                        forCommandID: ContentView.commandPaletteRightSidebarModeCommandID(mode)
+                    ),
+                    mode.shortcutAction
+                )
+            }
+        }
+    }
+
+    private func withSavedBetaFeatureDefaults(_ body: () throws -> Void) rethrows {
+        let defaults = UserDefaults.standard
+        let previousFeed = defaults.object(forKey: RightSidebarBetaFeatureSettings.feedEnabledKey)
+        let previousDock = defaults.object(forKey: RightSidebarBetaFeatureSettings.dockEnabledKey)
+        defer {
+            restore(previousFeed, forKey: RightSidebarBetaFeatureSettings.feedEnabledKey)
+            restore(previousDock, forKey: RightSidebarBetaFeatureSettings.dockEnabledKey)
+        }
+        try body()
+    }
+
+    private func restore(_ value: Any?, forKey key: String) {
+        let defaults = UserDefaults.standard
+        if let value {
+            defaults.set(value, forKey: key)
+        } else {
+            defaults.removeObject(forKey: key)
         }
     }
 }
